Move reader path selection helpers into utils
